fix(dap): pinned `DAPInstall` plugin commit.
It appears that the author of DAPInstall.nvim is rewriting this plugin,
but using the same repository on the main branch. As a result, we'll pin
to the last commit of `DAPInstall` as we wait for `dap-buddy` to mature.

References: Pocco81/dap-buddy.nvim#156e993516a49203b4b2094b3f06bf452ddcc5ec

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>